requirements.txt
django-countries==4.0
pinax-stripe==3.2.1
requests>=2.11.1
stripe==1.38.0
